refactor: rename BaseFhirPathAssertion to CollectionAssert and update assertion methods
